requirements.txt
discord.py
python-dotenv
rapidfuzz
yt-dlp
PyNaCl
